requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.10.0
websockets==12.0
aiofiles==23.2.1
jinja2==3.1.2
python-multipart==0.0.6

